+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,6 @@
     <string name="hello_2">Hello Test</string>
     <string name="submit">Submit</string>
     <string name="hello_test_label">Hello Test!</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
